backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
duckdb==0.10.0
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.7.0
faiss-cpu==1.8.0
bertopic==0.16.4
umap-learn==0.5.6
networkx==3.3.0
python-louvain==0.16
httpx==0.27.0
python-dotenv==1.0.1
groq==1.1.2
scikit-learn==1.4.2
hdbscan==0.8.42
gdown==4.7.3

